% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{init_obsnc}
\alias{init_obsnc}
\title{Initialize a new netCDF observation file}
\usage{
init_obsnc(fpath, stns, times, vnames)
}
\arguments{
\item{fpath}{File path for the new netCDF observation file}

\item{stns}{SpatialPointsDataFrame with station metadata}

\item{times}{POSIXct vector of observation times}

\item{vnames}{Vector of observation variable names (e.g.--tmin, tmax, prcp)}
}
\value{
A H5File object pointing to new netCDF observation file
}
\description{
Initialize a new netCDF observation file
}
