% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{load_obs}
\alias{load_obs}
\title{Load station observation time series as a "spacewide" xts object}
\usage{
load_obs(ds, vname, stns, times, start_end = NULL, stnids = NULL)
}
\arguments{
\item{ds}{H5File object pointing to netCDF observation file}

\item{vname}{Variable name for which to load observations}

\item{stns}{SpatialPointsDataFrame of station metadata from \code{\link{load_stns}}}

\item{times}{POSIXct vector of times for the netCDF observation file from
\code{\link{load_time}}}

\item{start_end}{Character vector of size 2 with start and end dates,
Default: loads observations for all times in the netCDF file}

\item{stnids}{Station ids for which to load observations,
Default: loads observations for all stations in the netCDF file}
}
\value{
A spacewide xts object of observations
}
\description{
Load station observation time series as a "spacewide" xts object
}
