% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{load_stnmeta}
\alias{load_stnmeta}
\title{Load station observation metadata from a csv file}
\usage{
load_stnmeta(fpath_csv, crs_string, ...)
}
\arguments{
\item{fpath_csv}{String pointing to a csv observation file}

\item{crs_string}{Coordinate reference system. Default: "+proj=longlat +datum=WGS84"}

\item{...}{Additional parameters that are passed to read.csv (e.g. skip)}
}
\value{
Station metadata as SpatialPointsDataFrame
}
\description{
Load station observation metadata from a csv file
}
