% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{load_stns}
\alias{load_stns}
\title{Load station metadata as a SpatialPointsDataFrame}
\usage{
load_stns(ds, vnames = c("station_id", "station_name", "longitude",
  "latitude", "elevation"), vname_stnid = "station_id",
  vnames_xy = c("longitude", "latitude"))
}
\arguments{
\item{ds}{H5File object pointing to netCDF observation file}

\item{vnames}{Variables names of station metadata to load,
Default: c("station_id", "station_name", "longitude", "latitude", "elevation")}

\item{vname_stnid}{Variable name from vnames representing unique station identifier,
Default: "station_id"}

\item{vnames_xy}{Variable names from vnames representing x,y position,
Default: c("longitude", "latitude")}
}
\value{
A SpatialPointsDataFrame
}
\description{
Load station metadata as a SpatialPointsDataFrame
}
