% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{load_tidy_obs}
\alias{load_tidy_obs}
\title{Load station observation time series from a csv file}
\usage{
load_tidy_obs(fpath_csv, ...)
}
\arguments{
\item{fpath_csv}{String pointing to a csv observation file}

\item{...}{Additional parameters that are passed to read.csv (e.g. skip)}
}
\value{
Observation time series as an xts object
}
\description{
Load station observation time series from a csv file
}
