% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{load_time}
\alias{load_time}
\title{Load time data as a vector of POSIXct times}
\usage{
load_time(ds, vname_time = "time")
}
\arguments{
\item{ds}{H5File object pointing to netCDF observation file}

\item{vname_time}{Variable name representing time, Default: "time"}
}
\value{
POSIXct vector of times
}
\description{
Load time data as a vector of POSIXct times
}
\seealso{
  \code{\link{load_obs}}, \code{\link{load_stns}}
}
\examples{
\dontrun{
# Set timezone for environment to UTC
Sys.setenv(TZ="UTC")

# Set file path for ChesWx netCDF observation file
PATH_CHESWX = "/Users/jaredwo/Downloads/data/cheswx"
path_obsnc <- file.path(PATH_CHESWX, "station_data/prcp_homog_19480101_20171231.nc")

# Open observation netCDF file
ds <- H5File$new(path_obsnc, mode='r')

# Load station metadata as a SpatialPointsDataFrame
stns <- load_stns(ds)

# Plot the station points
plot(stns)

# Load time variable data and create vector of POSIXct times
# This is needed when loading actual observations
times <- load_time(ds)

# Load precipitation time series for Richmond airport (station id: GHCND_USW00013740)
# Observations are returned as an xts object
ts_prcp <- load_obs(ds, 'prcp', stns, times, stnids=c('GHCND_USW00013740'))

# Plot Richmond time series
plot.xts(ts_prcp)

# Close netcdf file
ds$close_all()
}
}