% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{quick_load_stnobs}
\alias{quick_load_stnobs}
\title{Quickly load station metadata and observations from a netCDF observation file}
\usage{
quick_load_stnobs(fpath_ds, vname, start_end = NULL, stnids = NULL)
}
\arguments{
\item{fpath_ds}{File path to netCDF observation file}

\item{vname}{Variable name(s) for which to load observations}

\item{start_end}{Character vector of size 2 with start and end dates,
Default: loads observations for all times in the netCDF file}

\item{stnids}{Station ids for which to load observations,
Default: loads observations for all stations in the netCDF file}
}
\value{
List of station metadata and observations. Station metdata is a
  SpatialPointsDataFrame and observations are in a spacewise xts object.
}
\description{
Quickly load station metadata and observations from a netCDF observation file
}
\seealso{
  \code{\link{load_obs}}
}
\examples{
\dontrun{
# Set timezone for environment to UTC
Sys.setenv(TZ="UTC")

# Set file path for ChesWx netCDF observation file
PATH_CHESWX = "/Users/jaredwo/Downloads/data/cheswx"
path_obsnc <- file.path(PATH_CHESWX, "station_data/prcp_homog_19480101_20171231.nc")

# Load station metadata and observations in a single call
stnobs <- quick_load_stnobs(path_obsnc, 'prcp', c('1980-05-01','1980-05-31'))
stns <- stnobs[['stns']]
obs <- stnobs[['prcp']]
}
}