% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{build_pormask}
\alias{build_pormask}
\title{Build a boolean mask for stations that have long enough period-of-record}
\usage{
build_pormask(ds, vname, start_date, end_date, nyrs)
}
\arguments{
\item{ds}{H5File object pointing to netCDF observation file}

\item{vname}{Variable name for which to build the mask}

\item{start_date}{The start date for period-of-record time period}

\item{end_date}{The end date for period-of-record time period}

\item{nyrs}{The minimum period of record in years. The function tests
whether a station has at least nyrs years of data in each month.}
}
\value{
boolean mask for stations
}
\description{
Build a boolean mask for stations that have long enough period-of-record
}
