% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{coerce_spacewide_xts_to_STFDF}
\alias{coerce_spacewide_xts_to_STFDF}
\title{Convert a spacewide xts object to a spacetime STFDF object}
\usage{
coerce_spacewide_xts_to_STFDF(xts_sw, spdf_locs, varname)
}
\arguments{
\item{xts_sw}{A spacewide xts object.}

\item{spdf_locs}{Stations as a SpatialPointsDataFrame.}

\item{varname}{Variables names of station metadata to load.}
}
\value{
A STFDF object
}
\description{
Convert a spacewide xts object to a spacetime STFDF object
}
