% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{coerce_tidy_df_to_spacewide_xts}
\alias{coerce_tidy_df_to_spacewide_xts}
\title{Convert a tidy/longform dataframe into a spacewide xts object}
\usage{
coerce_tidy_df_to_spacewide_xts(df_tidy, value.var, time_col = "time",
  id_col = "station_id")
}
\arguments{
\item{df_tidy}{A tidy/longform dataframe.}

\item{value.var}{Name of column which stores values.}

\item{time_col}{Name of column which stores time. Default "time"}

\item{id_col}{Name of column which stores station id. Default: "station_id"}
}
\value{
A xts object
}
\description{
Convert a tidy/longform dataframe into a spacewide xts object
}
