% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{init_obsnc}
\alias{init_obsnc}
\title{Initialize a new netCDF observation file}
\usage{
init_obsnc(fpath, stns, times, vnames)
}
\arguments{
\item{fpath}{File path for the new netCDF observation file}

\item{stns}{SpatialPointsDataFrame with station metadata}

\item{times}{POSIXct vector of observation times}

\item{vnames}{Vector of observation variable names (e.g.--tmin, tmax, prcp)}
}
\value{
A H5File object pointing to new netCDF observation file
}
\description{
Initialize a new netCDF observation file
}
\seealso{
  \code{\link{quick_load_stnobs}}}
}
\examples{
\dontrun{
# Set timezone for environment to UTC
Sys.setenv(TZ="UTC")

# Set file path for ChesWx netCDF observation file
PATH_CHESWX = "/Users/jaredwo/Downloads/data/cheswx"
path_obsnc <- file.path(PATH_CHESWX, "station_data/prcp_homog_19480101_20171231.nc")

# Load station metadata and observations in a single call
stnobs <- quick_load_stnobs(path_obsnc, 'prcp', c('1980-05-01','1980-05-31'))
stns <- stnobs[['stns']]
obs <- stnobs[['prcp']]
times <- seq(as.Date('1980-05-01'),as.Date('1980-05-31'), by="day")

# Netcdf output file for point data
fpath_nc_out <- "prcp_homog_198005.nc"
ds_out <- init_obsnc(fpath_nc_out, stns, times, obs)
h5flush(ds_out)
h5close(ds_out)
}
}