% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_netcdf.R
\name{load_time}
\alias{load_time}
\title{Load time data as a vector of POSIXct times}
\usage{
load_time(ds, vname_time = "time")
}
\arguments{
\item{ds}{H5File object pointing to netCDF observation file}

\item{vname_time}{Variable name representing time, Default: "time"}
}
\value{
POSIXct vector of times
}
\description{
Load time data as a vector of POSIXct times
}
